+++ CHANGELOG.md
@@ -3182,3 +3182,4 @@
 - [2026-07-27 13:00] chore(storage): update dependencies and cleanup unused imports in storage
 - [2026-07-27 13:22] chore(ui): update dependencies and cleanup unused imports in ui
 - [2026-07-27 14:06] feat(network): implement shimmer loading skeleton in network
+- [2026-07-27 14:22] refactor(test): extract reusable atomic widgets from test